Make websocket connection upgrade header check case-insensitive
I was having troubles configuring an nginx reverse-proxy when
using Websockets.

It turns out my configuration was sending a "Connection: upgrade"
header, but Lotus expected "Connection: Upgrade".

This commit converts the check to be case-insensitive.

Some of the examples on the MDN page show lower-case "upgrade",
so I think it's not unusual for the usage to vary.